+++ packages/amap-types/map/layer/TileLayer.d.ts
@@ -97,5 +97,15 @@ declare namespace AMap {
      * 获取图层参数信息
      */
     getOptions(): object;
+
+    /**
+     * 卫星图层
+     */
+    static Satellite: typeof AMap.Satellite;
+
+    /**
+     * 实时交通图层
+     */
+    static Traffic: typeof AMap.Traffic;
   }
 }